+++ packages/shopify-app-remix/docs/staticPages/examples/index/splat-route.example.ts
@@ -1,0 +1,11 @@
+import {LoaderArgs} from '@remix-run/node';
+
+import shopify from '~/shopify.server';
+
+export async function loader({request}: LoaderArgs) {
+  await shopify.authenticate.admin(request);
+
+  // App logic goes here
+
+  return null;
+}