+++ packages/shopify-app-remix/src/server/authenticate/admin/authenticate.ts
@@ -497,20 +497,19 @@ export class AuthStrategy<
   }
 
   private redirectToBouncePage(url: URL): never {
-    const {api, config} = this;
+    const {config} = this;
 
-    // eslint-disable-next-line no-warning-comments
-    // TODO this is to work around a remix bug
-    // https://github.com/orgs/Shopify/projects/6899/views/1?pane=issue&itemId=28376650
-    url.protocol = `${api.config.hostScheme}:`;
-
-    const params = new URLSearchParams(url.search);
-    params.set('shopify-reload', url.href);
+    // Make sure we always point to the configured app URL so it also works behind reverse proxies (that alter the Host
+    // header).
+    url.searchParams.set(
+      'shopify-reload',
+      `${config.appUrl}${url.pathname}${url.search}`,
+    );
 
     // eslint-disable-next-line no-warning-comments
     // TODO Make sure this works on chrome without a tunnel (weird HTTPS redirect issue)
     // https://github.com/orgs/Shopify/projects/6899/views/1?pane=issue&itemId=28376650
-    throw redirect(`${config.auth.patchSessionTokenPath}?${params.toString()}`);
+    throw redirect(`${config.auth.patchSessionTokenPath}${url.search}`);
   }
 
   private createBillingContext(